% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Summary and detail of accuracy measures}
\usage{
accuracy(
  y_true,
  y_pred,
  y_bench,
  upper,
  lower,
  alpha,
  in_sample,
  frequency,
  error_detail_functions = c("ae", "se", "ape", "sape", "rae", "ase", "sase"),
  error_summary_functions = c("mae", "mse", "gmae", "mape", "smape", "mrae", "mdrae",
    "gmrae", "mase", "smase", "msis")
)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{in_sample}{in sample (hystorical values) of time series}

\item{frequency}{frequency of time series}
}
\value{
list with summary and details of accuracy measures
}
\description{
Summary and detail of accuracy measures
}
