% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{accuracy_detail}
\alias{accuracy_detail}
\title{Details (horizon specific) accuracy measures}
\usage{
accuracy_detail(
  y_true,
  y_pred,
  y_bench,
  in_sample,
  frequency,
  error_detail_functions = c("ae", "se", "ape", "sape", "rae", "ase", "sase")
)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{in_sample}{in sample (hystorical values) of time series}

\item{frequency}{frequency of time series}
}
\value{
horizon specific accuracy measures
}
\description{
Details (horizon specific) accuracy measures
}
