% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{accuracy_summary}
\alias{accuracy_summary}
\title{Summary of accuracy measures}
\usage{
accuracy_summary(
  y_true,
  y_pred,
  y_bench,
  in_sample,
  upper,
  lower,
  alpha,
  frequency,
  error_summary_functions = c("mae", "mse", "gmae", "mape", "smape", "mrae", "mdrae",
    "gmrae", "mase", "smase", "msis"),
  ...
)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{in_sample}{in sample (hystorical values) of time series}

\item{frequency}{frequency of time series}
}
\value{
summary and details of measures
}
\description{
Summary of accuracy measures
}
