% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{ae}
\alias{ae}
\title{Absolut error (AE)}
\usage{
ae(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns a vector of absolute differences
}
\description{
Absolut error (AE)
}
