% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{ape}
\alias{ape}
\title{Absolute percentage error (APE)}
\usage{
ape(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns vector of absolute percentage errors
}
\description{
Absolute percentage error (APE)
}
