% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{ase}
\alias{ase}
\title{Absolut scaled error (ASE)}
\usage{
ase(y_true, y_pred, in_sample, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{in_sample}{in sample (hystorical values) of time series}
}
\value{
Returns the absolute differences scaled by the mean in-sample error of naive
}
\description{
Absolut scaled error (ASE)
}
