% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{auto_damped}
\alias{auto_damped}
\title{Auto Damped Exponentional Smoothing}
\usage{
auto_damped(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
Auto Damped Exponentional Smoothing
}
