% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{auto_ses}
\alias{auto_ses}
\title{Auto Simple Exponential Smoothing}
\usage{
auto_ses(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
Auto Simple Exponential Smoothing
}
