% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{auto_snaive}
\alias{auto_snaive}
\title{Auto seasonal naive}
\usage{
auto_snaive(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
Auto seasonal naive
}
