% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{boxcox}
\alias{boxcox}
\title{BoxCox transformation}
\usage{
boxcox(y, ...)
}
\arguments{
\item{y}{time series vector}

\item{...}{not used}
}
\value{
list with two members; res is the transformed time series and param the used parameters for the transformation.
}
\description{
BoxCox transformation wrapper around the function \code{forecast::BoxCox()}.
}
