% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{combine_theta_lines}
\alias{combine_theta_lines}
\title{Theta forecasts}
\usage{
combine_theta_lines(y, h, combination, tl_zero, tl_two, theta)
}
\arguments{
\item{y}{historical observation in form of a time series object}

\item{h}{number of horizons to predict}

\item{combination}{combination type}

\item{tl_zero}{theta line zero}

\item{tl_two}{theta line two}

\item{theta}{theta parameter}
}
\value{
point forecasts including one prediction interval
}
\description{
Theta forecasts
}
