% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helper.R
\name{date_increment}
\alias{date_increment}
\title{Increment for date sequence}
\usage{
date_increment(y, direction = "future", ...)
}
\arguments{
\item{y}{teime series}

\item{direction}{direction of the increment}

\item{...}{not used}
}
\value{
increment
}
\description{
To be improved or removed
}
