% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helper.R
\name{decimal_to_date}
\alias{decimal_to_date}
\title{Decimal to date}
\usage{
decimal_to_date(dec_date, ...)
}
\arguments{
\item{dec_date}{decimal date}

\item{...}{not used}
}
\value{
date
}
\description{
Decimal to date
}
