% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{diff}
\alias{diff}
\title{Differencing of a vector}
\usage{
diff(y, n.diffs = 1, auto = T, ...)
}
\arguments{
\item{y}{vector which is correctly ordered with regards to time}

\item{n.diffs}{integer (default: 1) of differences to be conducted}

\item{auto}{boolean (default: T), whether number of differences should be defined automatically}

\item{...}{}
}
\value{
list with two members (res and param); where res is the
preporcessed vector and param the used parameters
}
\description{
Calculates the differences of time series over time.
}
