% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{.rx_forecast}
\alias{.rx_forecast}
\title{Wrapper to forecast with RevoscaleR}
\usage{
.rx_forecast(
  keys,
  data,
  h,
  levels,
  methods,
  pp_methods,
  point_combination,
  pi_combination_upper,
  pi_combination_lower,
  pool_limit,
  error_fun,
  weight_fun,
  val_h,
  sov_only,
  max_years,
  val_min_years,
  cv_min_years,
  cv_max_samples,
  allow_negatives,
  ...
)
}
\arguments{
\item{keys}{key of the time series, integer}

\item{data}{data of the time series passed by rxExecBy}

\item{params}{function parameters by rxExecBy}
}
\value{
forunco object
}
\description{
Create a wrapper around forunco that takes data from sql server or spark and 
predicts the future timesteps with forunco logic.
}
