% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{exp_inverse}
\alias{exp_inverse}
\title{Calculates the exponential inverse}
\usage{
exp_inverse(x, ...)
}
\arguments{
\item{x}{error metrics (higher is worse)}

\item{...}{not used}
}
\value{
the weights
}
\description{
Calculates the exponential inverse
}
