% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{g_mean}
\alias{g_mean}
\title{Geometric Mean}
\usage{
g_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector to take geometric mean from}

\item{na.rm}{Boolean (Optional, default: True) Wheater to remove NA values.}
}
\value{
geometric mean of vector
}
\description{
Geometric Mean
}
