% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{generic_combine}
\alias{generic_combine}
\title{Generic combination function}
\usage{
generic_combine(y_hat, c_fun, ...)
}
\arguments{
\item{y_hat}{predicted values for test set}

\item{...}{}

\item{c.fun}{combination funciton}
}
\value{
combined forecasts
}
\description{
This function is used for simpler combination methods supported out of the box from the stats package in R; such as e.g. \code{mean} or \code{median}
}
