% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{generic_forecast}
\alias{generic_forecast}
\title{Wrapper to preprecess, predict and postprocess forecasts}
\usage{
generic_forecast(
  fcs.fun,
  y,
  h,
  level,
  pp.operations = c("seasonal_adjustment"),
  ...
)
}
\arguments{
\item{fcs.fun}{forecasting function}

\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{pp.operations}{preprocessing operations to be conducted}

\item{...}{not used}
}
\value{
forecast object
}
\description{
Wrapper to preprecess, predict and postprocess forecasts
}
