% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{gmae}
\alias{gmae}
\title{Geomatric mean absolute error (GMAE)}
\usage{
gmae(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns geomatric mean of squared errors
}
\description{
Geomatric mean absolute error (GMAE)
}
