% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{gmrae}
\alias{gmrae}
\title{Geomatric mean relative absolut error (GMRAE)}
\usage{
gmrae(y_true, y_pred, y_bench, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{y_bench}{forecasts of benchmark method}
}
\value{
Returns geometric mean of relative absolute differences
}
\description{
Geomatric mean relative absolut error (GMRAE)
}
