% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_boxcox}
\alias{inv_boxcox}
\title{Inverse Boxcox transformation}
\usage{
inv_boxcox(y, param, ...)
}
\arguments{
\item{y}{time series}

\item{param}{parameter list}

\item{...}{not used}
}
\value{
backtransformed time series vector
}
\description{
Inverses the BoxCox transformation conducted in function \code{boxcox}. The funciton is a wrapper around the function \code{forecast::InvBoxCox()}.
}
