% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_diff}
\alias{inv_diff}
\title{Inverses diff transformations}
\usage{
inv_diff(y, param, is_fcs = T, ...)
}
\arguments{
\item{y}{time series vector}

\item{param}{parameter of the transformation function}

\item{is_fcs}{boolean, default (T). whether it is a forecast or in-sample vector.}

\item{...}{not used}
}
\value{
backtransformed time series
}
\description{
Backtransformes diff transformations performed in \code{diff()}. Support for multiple transformations.
}
