% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_log}
\alias{inv_log}
\title{Inverse of log transformation}
\usage{
inv_log(y, ...)
}
\arguments{
\item{y}{time series vector.}

\item{...}{not used}
}
\value{
backtransformed time series
}
\description{
Used to automatically perfrom multiply transformations using the preprocessor environment; basicall calculates \code{exp} of input \code{y}.
}
