% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_normalize}
\alias{inv_normalize}
\title{Inverses normalization}
\usage{
inv_normalize(y, param = NA, ...)
}
\arguments{
\item{y}{vector to be rescaled}

\item{param}{parameters used for the original scaling}

\item{...}{}
}
\value{
renormalized vector
}
\description{
Reverses the normalization conducted in function \code{normalize()}.
}
