% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_scale}
\alias{inv_scale}
\title{Inverses scaling}
\usage{
inv_scale(y, param = NA, ...)
}
\arguments{
\item{y}{vector to be rescaled}

\item{param}{parameters used for the original scaling}

\item{...}{}
}
\value{
rescaled vector
}
\description{
Reverses the scaling conducted in function \code{scale()}.
}
