% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{inv_seasonal_adjustment}
\alias{inv_seasonal_adjustment}
\title{Inversed seasonal adjustment}
\usage{
inv_seasonal_adjustment(y, param, ...)
}
\arguments{
\item{y}{time series}

\item{param}{parameter list}

\item{...}{not used}
}
\value{
Re-seasonalized time series
}
\description{
Backtransforms the seasonal adjustment conducted in function \code{seasonal_adjustment}.
}
\references{
Makridakis SG, Wheelwright SC, Hyndman RJ (1998). Forecasting: Methods and applications ( Third Edition). New York: Wiley.
}
