% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{inverse}
\alias{inverse}
\title{Calculates the inverse}
\usage{
inverse(x, ...)
}
\arguments{
\item{x}{error metrics (higher is worse)}

\item{...}{not used}
}
\value{
the weights
}
\description{
Calculates the inverse
}
