% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{is_seasonal}
\alias{is_seasonal}
\title{Seasonality test}
\usage{
is_seasonal(y, frequency = frequency(y), ...)
}
\arguments{
\item{y}{time series}

\item{frequency}{frequency of time series vector}

\item{...}{not used}
}
\value{
boolean (true if seasonal, false otherwise)
}
\description{
Seasonality test
}
\references{
Makridakis SG, Wheelwright SC, Hyndman RJ (1998). Forecasting: Methods and applications ( Third Edition). New York: Wiley.
}
