% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mae}
\alias{mae}
\title{Mean absolute error (MAE)}
\usage{
mae(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns ean of absolute differences
}
\description{
Mean absolute error (MAE)
}
