% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mape}
\alias{mape}
\title{Mean absolute percentage error (MAPE)}
\usage{
mape(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns mean of absolute percentage differences
}
\description{
Mean absolute percentage error (MAPE)
}
