% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mase}
\alias{mase}
\title{Mean absolut scaled error (MASE)}
\usage{
mase(y_true, y_pred, in_sample, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{in_sample}{in sample (hystorical values) of time series}
}
\value{
Returns mean of the absolute differences scaled by the mean in-sample error of naive
}
\description{
Mean absolut scaled error (MASE)
}
