% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mdrae}
\alias{mdrae}
\title{Median relative absolut error (MRAE)}
\usage{
mdrae(y_true, y_pred, y_bench, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{y_bench}{forecasts of benchmark method}
}
\value{
Returns median of relative absolute differences
}
\description{
Median relative absolut error (MRAE)
}
