% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mrae}
\alias{mrae}
\title{Mean relative absolut error (MRAE)}
\usage{
mrae(y_true, y_pred, y_bench, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{y_bench}{forecasts of benchmark method}
}
\value{
Returns mean of relative absolute differences
}
\description{
Mean relative absolut error (MRAE)
}
