% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{mse}
\alias{mse}
\title{Mean square error (MSE)}
\usage{
mse(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns mean of squared errors
}
\description{
Mean square error (MSE)
}
