% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{msis}
\alias{msis}
\title{Mean Scaled interval score (MSIS)}
\usage{
msis(y_true, upper, lower, alpha, in_sample, frequency, ...)
}
\arguments{
\item{y_true}{true values of forecasts}

\item{upper}{upper bound}

\item{lower}{lower bound}

\item{alpha}{significance level}

\item{in_sample}{in_sample data}

\item{frequency}{frequency of data}

\item{...}{additional parameters, not used}
}
\value{
Average MSIS
}
\description{
Mean Scaled interval score (MSIS)
}
