% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{normalize}
\alias{normalize}
\title{Normalize vector}
\usage{
normalize(y, ...)
}
\arguments{
\item{y}{vector of occurences}

\item{...}{}
}
\value{
list with two members (res and param); where res is the
preporcessed vector and param the used parameters
}
\description{
Normalize vector
}
