% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{plot_forunco}
\alias{plot_forunco}
\title{Plot of forunco object}
\usage{
plot_forunco(forunco, interactive = F, future_vals = NA, ...)
}
\arguments{
\item{forunco}{forunco object}

\item{interactive}{whether or not to return an interactive plot or a static one}

\item{...}{addtional argument, currently not used}

\item{add_data}{additional data}
}
\value{
ggplot/plotly object
}
\description{
Plot of forunco object
}
