% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{pool_mean}
\alias{pool_mean}
\title{Calculates the mean of the pool}
\usage{
pool_mean(x, ...)
}
\arguments{
\item{x}{error metrics (higher is worse)}

\item{...}{not used}
}
\value{
the weights
}
\description{
Calculates the mean of the pool
}
