% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{pool_median}
\alias{pool_median}
\title{Calculates the mean of the pool}
\usage{
pool_median(fcs, ...)
}
\arguments{
\item{fcs}{pooled forecasts}

\item{...}{not used}
}
\value{
the weights
}
\description{
Calculates the mean of the pool
}
