% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{preprocessor}
\alias{preprocessor}
\title{Preprocessor environment}
\usage{
preprocessor()
}
\value{
returns either transformed or backtransformed time series vector
}
\description{
Can be used to flexibly transforming and backtransforming data. Generally all functions - including custom ones in the current environment are supported.
The function assumes that each transformation funcotion (e.g. \code{normalize <- function (x, ...) {}}) has a backtransformation function with the same name
but a \code{inv_} prefix (e.g. \code{inv_normalize <- function(x, param, ...){}}). All parameter used for transforming the variables can be stored in the
param variable; which will be passed automatically back to the backtransformation function.
}
\details{
The preprocessor has to "instantiated" and can then be called using the following parameters:
\itemize{
 \item{"y"}{time series or forecast to be transformed/backtransformed}
 \item{"action"}{action to be taked, either 'transform' or 'backtransform'}
 \item{"operations"}{transformatoin operations to be takes (functions to be called)}
 \item{"n.diffs"}{integer default (NA). Sets the number of differencing to be taken (used only when diff is part of operator)}
 \item{"auto"}{boolean, default (T), whether the number of differencing is to be determined automatically}
 \item{"is_fcs"}{boolean ,default (T), wheter it is forecast or in-sample data to be transformed}
 \item{"frequency"}{inteher, default (NA), the frequency of the time series (used for seasonal adjustment)}
 \item{"h"}{horizon, default (1)}
}
}
\examples{
# not run
# load demo data
library(Mcomp)
# get demo time series
demo_ts <- M3[[1560]]$x
# show demo ts
demo_ts
# get preprocessor "object" (environment)
pp <- preprocessor()
# preprocess time series
y_pp <- pp(y = demo_ts,
           action = 'transform',
           operations = c('seasonal_adjustment', 'scale'),
           frequency = frequency(demo_ts))
# show preprocessed time series
y_pp
# ok, now go back to the roots
y_back <- pp(y = y_pp, action = 'backtransform', is_fcs = F)
y_back - demo_ts
# end not run
}
