% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_combination.R
\name{produce_forecasts}
\alias{produce_forecasts}
\title{Produces forecasts}
\usage{
produce_forecasts(
  y,
  h,
  levels = c(80, 95),
  methods = c("auto_ets", "auto_arima", "auto_thetaf"),
  ...
)
}
\arguments{
\item{y}{Time series object.}

\item{h}{Horizons to be predicted.}

\item{levels}{Prediction interval levels. Optional, default \code{c(95)}.}

\item{methods}{Methods to be combined. Optional, default \code{auto_ets, auto_arima, auto_thetaf}.}

\item{...}{passed to the forecasting functions}

\item{pp_obj}{preprocessing parameters}
}
\value{
combined forecasts of time series y including confidence intervals
}
\description{
Using the predefined methods, this function produces point forecasts as well as prediction intervals of any given level.
}
