% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{rae}
\alias{rae}
\title{Relative absolut error (RAE)}
\usage{
rae(y_true, y_pred, y_bench, epsilon = 0.001, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}

\item{y_bench}{forecasts of benchmark method}
}
\value{
Returns a vector of relative absolute differences
}
\description{
Relative absolut error (RAE)
}
