% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replacement of outliers}
\usage{
replace_outliers(y, frequency, start = 1, ...)
}
\arguments{
\item{y}{time series vector}

\item{frequency}{frequency of time series vector}

\item{start}{start date, default (1)}

\item{...}{not used}
}
\value{
timeseries without outliers
}
\description{
This function is a wrapper around the funciton \code{forecast::tsoutliers()}.
}
