% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{rmse}
\alias{rmse}
\title{Root mean square error (MSE)}
\usage{
rmse(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns root mean of squared errors
}
\description{
Root mean square error (MSE)
}
