% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{sape}
\alias{sape}
\title{Symmetric absolute percentage error (sAPE)}
\usage{
sape(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns vector of symmetric absolute percentage differences
}
\description{
Symmetric absolute percentage error (sAPE)
}
