% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{se}
\alias{se}
\title{Squared error (SE)}
\usage{
se(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns a vector of saured differences
}
\description{
Squared error (SE)
}
