% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_preprocess.R
\name{seasonal_adjustment}
\alias{seasonal_adjustment}
\title{Conductes a seasonal adjustment}
\usage{
seasonal_adjustment(y, frequency = frequency, h = h, type = "M", ...)
}
\arguments{
\item{y}{time series vector}

\item{frequency}{frequency of time series}

\item{h}{horizon of time series}

\item{...}{not used}
}
\value{
list with two members; res is the seasonally adjusted time series; param the seasonal components to reconstruct the time series or forecasts in \code{inv_seasonal_adjustment}.
}
\description{
Seasonally adjusts a time series vector using the multiplicative decomposition approach.
}
