% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{shd}
\alias{shd}
\title{Conventional SHD Combination}
\usage{
shd(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
Conventional SHD Combination
}
