% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_combination.R
\name{simple_combination}
\alias{simple_combination}
\title{Simple combination of univariate time series forecasting methods}
\usage{
simple_combination(
  y,
  h,
  fcs_mats = NA,
  levels,
  methods,
  point_combination = "median",
  pi_combination_upper = "median",
  pi_combination_lower = "median",
  allow_negatives = F,
  ...
)
}
\arguments{
\item{y}{Time series object.}

\item{h}{Horizons to be predicted.}

\item{levels}{Prediction interval levels. Optional, default \code{c(95)}.}

\item{methods}{Methods to be combined. Optional, default \code{auto_ets, auto_arima, auto_thetaf}.}

\item{point_combination}{Point forecast combination operator. Optional, default \code{meidan}.}

\item{pi_combination_upper}{combination operator of the upper bound of the prediction intervals. Optional, default \code{max}.}

\item{pi_combination_lower}{combination operator for the lower bounds of the prediction intervals. Optional, default \code{min}.}

\item{allow_negatives}{Flag indicating whether to allow negative values or not. Optional, default \code{FALSE}.}

\item{...}{passed to the forecasting functions}
}
\value{
combined forecasts of time series y including confidence intervals
}
\description{
Combines muultiple forecasts using definable combination operators.
}
