% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_accuracy.R
\name{smape}
\alias{smape}
\title{Symmetric mean absolute percentage error (sMAPE)}
\usage{
smape(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{actual value of time series}

\item{y_pred}{predicted value of time series}
}
\value{
Returns mean of symmetric absolute percentage differences
}
\description{
Symmetric mean absolute percentage error (sMAPE)
}
