% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{squared_inverse}
\alias{squared_inverse}
\title{Calculates the squared inverse}
\usage{
squared_inverse(x, ...)
}
\arguments{
\item{x}{error metrics (higher is worse)}

\item{...}{not used}
}
\value{
the weights
}
\description{
Calculates the squared inverse
}
