% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_fit}
\alias{theta_fit}
\title{Theta fit}
\usage{
theta_fit(y, h, level = 95, theta, curve, combination, seasonality)
}
\arguments{
\item{y}{historical observation in form of a time series object}

\item{h}{number of horizons to predict}

\item{level}{prediction interval level, only one allowed.}

\item{theta}{theta parameter.}

\item{curve}{curve type.}

\item{seasonality}{seasonality type.}

\item{model}{model type.}
}
\value{
point forecasts including one prediction interval level
}
\description{
This method is inspired by the implementation of E. Spiliotis and V. Assimakopoulos (2017) / Forecasting & Strategy Unit - NTUA.
}
