% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_forecast}
\alias{theta_forecast}
\title{theta forecast}
\usage{
theta_forecast(y, h, level = 95, model = "MEM")
}
\arguments{
\item{y}{historical observations}

\item{h}{horizons to be predicted}

\item{level}{prediction interval level}

\item{model}{model to be predicted}
}
\value{
forecast object
}
\description{
theta forecast
}
