% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_line}
\alias{theta_line}
\title{Theta line}
\usage{
theta_line(y, tl_zero, theta, combination)
}
\arguments{
\item{y}{historical observarions}

\item{tl_zero}{theta line zero}

\item{theta}{theta parameter}

\item{combination}{combination type}
}
\value{
theta line to be fitted by ses
}
\description{
Theta line
}
