% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_line_zero}
\alias{theta_line_zero}
\title{Theta line zero}
\usage{
theta_line_zero(y, h, curve, level, ...)
}
\arguments{
\item{y}{historical observations}

\item{h}{horizons to predict}

\item{curve}{curve type}

\item{level}{prediction interval level}

\item{...}{}
}
\value{
thea line zero
}
\description{
Theta line zero
}
