% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_mem}
\alias{theta_mem}
\title{MEM theta model}
\usage{
theta_mem(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
MEM theta model
}
