% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_predictors.R
\name{theta_mla}
\alias{theta_mla}
\title{MLA theta model}
\usage{
theta_mla(y, h, level, ...)
}
\arguments{
\item{y}{historical values}

\item{h}{number of horizons to predict}

\item{level}{prediction interval levels}

\item{...}{not used}
}
\value{
forecast object
}
\description{
MLA theta model
}
