% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helper.R
\name{tidier_ts}
\alias{tidier_ts}
\title{Tidy time series in data frame}
\usage{
tidier_ts(y, type = "actual", ...)
}
\arguments{
\item{y}{observations}

\item{type}{type of observations}

\item{...}{not used}
}
\value{
tidy ts
}
\description{
Tidy time series in data frame
}
