% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helper.R
\name{time_sequence}
\alias{time_sequence}
\title{Time sequence}
\usage{
time_sequence(y, start_pos = "max", out_length)
}
\arguments{
\item{y}{observations}

\item{start_pos}{where to start}

\item{out_length}{how many to generate}
}
\value{
time sequence
}
\description{
To be improved or removed in the future.
}
