% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{weight_error}
\alias{weight_error}
\title{Weighting and pooling methods on the basis of their error}
\usage{
weight_error(weight_fun, error, y_hat, pool_limit, ...)
}
\arguments{
\item{weight_fun}{weight function}

\item{error}{the error vector which should be weightened}

\item{pool_limit}{the pool size to be considered (e.g. 3 would indicate that only the best three methods would be selected)}

\item{...}{passed to weight_fun}
}
\value{
the weights
}
\description{
Weighting and pooling methods on the basis of their error
}
