% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_operators.R
\name{weighted_average}
\alias{weighted_average}
\title{Averaging forecasts using weights}
\usage{
weighted_average(
  val_hat,
  val_true,
  y_hat,
  error_fun = "rmse",
  weight_fun = "inverse",
  pool_limit = 3,
  ...
)
}
\arguments{
\item{val_hat}{forcasts on the validation set}

\item{val_true}{true values of the validation set}

\item{y_hat}{forecasts for the test set}

\item{weight_fun}{the function to determine the weights}

\item{pool_limit}{how many methods should be considered for combination}

\item{...}{not used}

\item{error_measure}{error measure to be used for error calculation; error measures that calculate errors per horizons will lead to a horizon specific combination, error measures over all horizons on the other hand, will lead to a simple weighted combination.}
}
\value{
combined forecasts using imrse for weighting
}
\description{
Combines forecasting methods based on the validation error. See Nowotarski, J., Raviv, E., Tr\"uck, S., and Weron, R. (2014) for more examples.
}
\references{
Nowotarski, J., Raviv, E., Tr\"uck, S., and Weron, R. (2014). An Empirical Comparison of Alternative
Schemes for Combining Electricity Spot Price Forecasts. \emph{Energy Economics}, \bold{46}, 395--412.
}
